open-logic-bit: separate scan from device open
The current OLB driver sets up the ftdi context on scan, then performs
device init on open. The scan only supports one ftdi endpoint, which
breaks if the first FTDI endpoint discovered is not an OLB device.

Instead, change the scan callback to do an actual scan, setting up the
sdi->connection_id for discovered FTDI devices. This allows the correct
device to be selected, then will be properly initialised in the open
callback.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>